created confirm dialog to confirm delete

+++ src/components/ListItem.jsx
@@ -22,7 +22,11 @@ export function ListItem({ name, data, listToken }) {
 	};
 
 	function handleDeleteItem() {
-		deleteItem(listToken, data.id);
+		if (window.confirm(`Are you sure you wish to delete ${name}?`)) {
+			deleteItem(listToken, data.id);
+		} else {
+			this.onCancel();
+		}
 	}
 
 	return (
